♻️ refactor(smithers): share prepare output schemas
Relates to #307

Extract duplicated prepareOutputSchemas logic from create.js and
create-external-smithers.js into a shared module at
packages/smithers/src/prepareOutputSchemas.js. Both callers now import
the extracted function, eliminating ~60 lines of duplication. Tests in
create-unit.test.js updated to cover the shared path.
